refactor: extracted random group formation function into groupControllerHelpers.js
